mods per @rakuy0
