+++ vivisect/tests/helpers.py
@@ -4,8 +4,6 @@ import functools
 
 import vivisect.cli as v_cli
 
-import gc
-gc.set_threshold(500, 8, 8)
 
 class MockVw(object):
     def __init__(self, *args, **kwargs):
